Trait vectils::Vector2d

pub trait Vector2d<T> where
    T: PartialEq + PartialOrd + Copy + Clone + Add<Output = T> + Sub<Output = T> + Mul<Output = T> + Div<Output = T>, {
    fn xi(&self) -> T;

    fn yi(&self) -> T;

    fn xf(&self) -> T;

    fn yf(&self) -> T;

    fn in_range(a: T, b: T, c: T) -> bool { ... }

    fn w(&self) -> T { ... }

    fn h(&self) -> T { ... }

    fn delta(&self) -> T { ... }

    fn in_dom(&self, pt_x: T) -> bool { ... }

    fn linear(&self, pt_x: T) -> T { ... }

    fn cross<O: Vector2d<T>>(&self, b: &O) -> Option<[T; 2]> { ... }
}

This trait allows the creation of 2 dimensional vectors over various types.

Required Methods

fn xi(&self) -> T

X value of first point.

fn yi(&self) -> T

Y value of first point.

fn xf(&self) -> T

X value of second point.

fn yf(&self) -> T

Y value of second point.

Provided Methods

fn in_range(a: T, b: T, c: T) -> bool

Associated helper function which checks for whether or not a number falls within a range (inclusive).

fn w(&self) -> T

Returns width of the vector. Signed.

fn h(&self) -> T

Returns height of the vector. Signed.

fn delta(&self) -> T

Returns the slope of the vector.

fn in_dom(&self, pt_x: T) -> bool

Returns whether or not this vector's domain includes the x value provided.

fn linear(&self, pt_x: T) -> T

Returns the y value of the line which is made by extending the ends of this vector.

fn cross<O: Vector2d<T>>(&self, b: &O) -> Option<[T; 2]>

Returns the point at which two vectors cross, or None if they do not.